services/backend/serenitys-keys-backend/requirements.txt
aiosqlite==0.19.0
alembic==1.12.0
apscheduler==3.10.4
asyncpg==0.28.0
fastapi==0.103.2
google-api-python-client==2.95.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
httpx==0.25.0
pydantic==2.3.0
pydantic-settings==2.0.2
PyJWT==2.8.0
python-dateutil==2.8.2
sentry-sdk==1.31.0
slowapi==0.1.8
sqlalchemy==2.0.21
stripe==5.4.0
uvicorn[standard]==0.23.2
